i#4418 vmheap: Change sized options to use pointer-sized types
Changes the "uint_size" option type to be pointer-sized, to better
handle larger values we are now wanting for things like -vmheap_size,
especially on ARM machines with larger page sizes.

Adds some sanity tests to ensure all the parsing and printing code was
updated.

Issue: #4418

+++ core/module_shared.h
@@ -464,7 +464,7 @@ typedef struct {
 void
 module_calculate_digest(/*OUT*/ module_digest_t *digest, app_pc module_base,
                         size_t module_size, bool full_digest, bool short_digest,
-                        uint short_digest_size, uint sec_char_include,
+                        size_t short_digest_size, uint sec_char_include,
                         uint sec_char_exclude);
 
 /* actually in utils.c since os-independent */
